requirements.txt
streamlit
pypdf
python-dotenv
google-generativeai
